+++ conf_testing/rules/test_parameter_files.py
@@ -7,7 +7,8 @@ log = logging.getLogger('HABApp.TestParameterFiles')
 
 # User Parameter files to create rules dynamically
 try:
-    assert HABApp.parameters.get_parameter_value('param_file', 'key') != 10, f'Loading of Parameters does not work properly'
+    assert HABApp.Parameters.get_parameter_value('param_file', 'key') != 10, \
+        f'Loading of Parameters does not work properly'
 except Exception as e:
     log.error(e)
 
@@ -21,7 +22,7 @@ class TestParamFile(TestBaseRule):
         self.add_test('ParamFile', self.test_param_file)
 
     def test_param_file(self):
-        p = HABApp.parameters.get_parameter('param_file', 'key')
+        p = HABApp.Parameters.get_parameter('param_file', 'key')
         assert p < 11
         assert p.value == 10
         return True
